Adding new packages, ProductDbService, ProductMapper

+++ src/main/java/com/kodilla/ecommercee/domain/ProductDto.java
@@ -10,6 +10,6 @@ public class ProductDto {
 
     int productId;
     String productName;
-    int productPrice;
+    double productPrice;
     String productDescription;
 }
